+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Feature
 * add forms extension [#610](https://github.com/jcubic/jquery.terminal/issues/610)
 * rewrite xml_formatter and add new tags
+* add animation interface
 ### Bugfix
 * fix rest of the extension files in NodeJS (regression since 2.22.0)
 * fix inheritance of attributes (including styles) and classes in nested formatting